Sneaky unrelated refactoring - move form to the correct module.
