+++ django_app/redbox_app/redbox_core/views.py
@@ -1,10 +1,8 @@
 import logging
 import uuid
-from collections.abc import Mapping, MutableSequence, Sequence
+from collections.abc import MutableSequence, Sequence
 from pathlib import Path
-from typing import ClassVar
 
-from django import forms
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import FieldError, ValidationError
@@ -18,6 +16,7 @@ from django.views import View
 from django.views.decorators.http import require_http_methods
 from django.views.generic import UpdateView
 from redbox_app.redbox_core.client import CoreApiClient
+from redbox_app.redbox_core.forms import DemographicsForm
 from redbox_app.redbox_core.models import (
     ChatHistory,
     ChatMessage,
@@ -322,22 +321,6 @@ class CheckDemographicsView(View):
             return redirect(documents_view)
         else:
             return redirect("demographics")
-
-
-class DemographicsForm(forms.ModelForm):
-    class Meta:
-        model = User
-        fields = ("business_unit", "grade", "profession")
-        labels: ClassVar[Mapping[str, str]] = {
-            "business_unit": "Business unit",
-            "grade": "Grade",
-            "profession": "Profession",
-        }
-        widgets: ClassVar[Mapping[str, forms.Widget]] = {
-            "business_unit": forms.Select(attrs={"class": "govuk-select"}),
-            "grade": forms.Select(attrs={"class": "govuk-select"}),
-            "profession": forms.Select(attrs={"class": "govuk-select"}),
-        }
 
 
 class DemographicsView(UpdateView):
